feat: fix benchmark bug and add personal rmp config
