+++ config/level2_PourLiquid.yaml
@@ -27,12 +27,12 @@ task:
   left_pos:
     x: [0.17, 0.32]
     y: [-0.35, -0.2]
-    z: [0.86, 0.86]
+    z: [0.87, 0.87]
   obj_paths:
     - path: "/World/beaker2"
       position_range:
         x: [0.22, 0.37]
-        y: [0.0, 0.15]
+        y: [0.07, 0.22]
         z: [0.9, 0.9]
