+++ controllers/base_controller.py
@@ -7,7 +7,7 @@ from controllers.robot_controllers.grapper_manager import Gripper
 from controllers.robot_controllers.trajectory_controller import FrankaTrajectoryController
 from factories.collector_factory import create_collector
 from utils.object_utils import ObjectUtils
-from isaacsim.robot.manipulators.examples.franka.controllers.rmpflow_controller import RMPFlowController as FrankaRMPFlowController
+from robots.franka.rmpflow_controller import RMPFlowController as FrankaRMPFlowController
 
 class BaseController(ABC):
     """Base class for all controllers in the chemistry lab simulator.
@@ -16,7 +16,7 @@ class BaseController(ABC):
     and episode tracking.
     """
     
-    def __init__(self, cfg, robot):
+    def __init__(self, cfg, robot, use_default_config=True):
         """Initialize the base controller.
         
         Args:
@@ -39,7 +39,8 @@ class BaseController(ABC):
         
         self.rmp_controller = FrankaRMPFlowController(
             name="target_follower_controller",
-            robot_articulation=robot
+            robot_articulation=robot,
+            use_default_config=use_default_config
         )
 
         self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
