+++ controllers/openclose_controller.py
@@ -1,7 +1,7 @@
 import re
 from typing import Optional
 from controllers.atomic_actions.close_controller import CloseController
-from isaacsim.robot.manipulators.examples.franka.controllers.rmpflow_controller import RMPFlowController
+from robots.franka.rmpflow_controller import RMPFlowController
 import numpy as np
 
 from controllers.atomic_actions.open_controller import OpenController
