+++ controllers/placepress_controller.py
@@ -1,4 +1,4 @@
-from isaacsim.robot.manipulators.examples.franka.controllers.rmpflow_controller import RMPFlowController
+from robots.franka.rmpflow_controller import RMPFlowController
 from scipy.spatial.transform import Rotation as R
 import numpy as np
 from enum import Enum
