+++ controllers/stir_controller.py
@@ -2,7 +2,7 @@ from typing import Optional
 import numpy as np
 
 from isaacsim.core.utils.types import ArticulationAction
-from isaacsim.robot.manipulators.examples.franka.controllers.rmpflow_controller import RMPFlowController
+from robots.franka.rmpflow_controller import RMPFlowController
 from scipy.spatial.transform import Rotation as R
 
 from .base_controller import BaseController
