+++ requirements.txt
@@ -9,19 +9,15 @@ zarr
 tqdm
 einops
 wandb
-pynput
-click
 threadpoolctl
 matplotlib
 dill
 pydantic
 rtree
-shapely
-open3d
 diffusers
 pytorch-lightning
 lightning-bolts
 tyro
-sentry-sdk==1.43.0
 typing_extensions==4.12.2
 numpy==1.26.0
+sentry-sdk==2.29.1
