+++ robots/franka/lula_franka_gen.urdf
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:e9024642e7952cbcaec0ae14425bf1cd19d674d3c98eeef3793913f63a8101b6
+size 14771
