+++ robots/franka/rmpflow/config_no_feedback.json
@@ -1,0 +1,10 @@
+{
+    "end_effector_frame_name": "right_gripper",
+    "maximum_substep_size" : 0.00167,
+    "ignore_robot_state_updates" : true,
+    "relative_asset_paths": {
+        "robot_description_path" : "robot_descriptor.yaml",
+        "urdf_path": "../lula_franka_gen.urdf",
+        "rmpflow_config_path" : "franka_rmpflow_common.yaml"
+    }
+}
