+++ robots/franka/rmpflow/franka_rmpflow_common.yaml
@@ -1,0 +1,115 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+joint_limit_buffers: [.01, .03, .01, .01, .01, .01, .01]
+
+rmp_params:
+    cspace_target_rmp:
+        metric_scalar: 50.
+        position_gain: 200.
+        damping_gain: 50.
+        robust_position_term_thresh: .5
+        inertia: 1.
+    cspace_trajectory_rmp:
+        p_gain: 100.
+        d_gain: 10.
+        ff_gain: .25
+        weight: 50.
+    cspace_affine_rmp:
+        final_handover_time_std_dev: .25
+        weight: 2000.
+    joint_limit_rmp:
+        metric_scalar: 1000.
+        metric_length_scale: .01
+        metric_exploder_eps: 1e-3
+        metric_velocity_gate_length_scale: .01
+        accel_damper_gain: 200.
+        accel_potential_gain: 1.
+        accel_potential_exploder_length_scale: .1
+        accel_potential_exploder_eps: 1e-2
+    joint_velocity_cap_rmp:
+        max_velocity: 4.  # max_xd
+        velocity_damping_region: 1.5
+        damping_gain: 1000.0
+        metric_weight: 100. # metric_scalar
+    target_rmp:
+        accel_p_gain: 50.
+        accel_d_gain: 85.
+        accel_norm_eps: .075
+        metric_alpha_length_scale: .05
+        min_metric_alpha: .01
+        max_metric_scalar: 10000
+        min_metric_scalar: 2500
+        proximity_metric_boost_scalar: 20.
+        proximity_metric_boost_length_scale: .02
+        xi_estimator_gate_std_dev: 20000.
+        accept_user_weights: false  # Values >= .5 are true and < .5 are false
+    axis_target_rmp:
+        accel_p_gain: 80.
+        accel_d_gain: 30.
+        metric_scalar: 2
+        proximity_metric_boost_scalar: 800.
+        proximity_metric_boost_length_scale: .08
+        xi_estimator_gate_std_dev: 20000.
+        accept_user_weights: false
+    collision_rmp:
+        damping_gain: 50.
+        damping_std_dev: .04
+        damping_robustness_eps: 1e-2
+        damping_velocity_gate_length_scale: .01
+        repulsion_gain: 800.
+        repulsion_std_dev: .01
+        metric_modulation_radius: .5
+        metric_scalar: 10000.  # Real value should be this.
+        #metric_scalar: 0.  # Turns off collision avoidance.
+        metric_exploder_std_dev: .02
+        metric_exploder_eps: .001
+    damping_rmp:
+        accel_d_gain: 30.
+        metric_scalar: 50.
+        inertia: 100.
+
+canonical_resolve:
+    max_acceleration_norm: 50.
+    projection_tolerance: .01
+    verbose: false
+
+body_cylinders:
+    - name: base_stem
+      pt1: [0,0,.333]
+      pt2: [0,0,0.]
+      radius: .05
+    - name: base_tee
+      pt1: [0,0,.333]
+      pt2: [0,0,.333]
+      radius: .15
+
+# Each arm is approx. 1m from (arm) base to gripper center.
+# .1661 between links (approx .15)
+body_collision_controllers:
+    - name: panda_link7
+      radius: .005
+    - name: panda_wrist_end_pt
+      radius: .005
+    - name: panda_hand
+      radius: .005
+    - name: panda_face_left
+      radius: .005
+    - name: panda_face_right
+      radius: .005
+    - name: panda_leftfingertip
+      radius: .0075
+    - name: panda_rightfingertip
+      radius: .0075
